monodromy abstraction
Adds the "MonodromyModel" type as an abstraction over the raw "lifting.jl" functions. It enforces more invariants, like ensuring the base path and all its lifts are the same length. Long-term, this is a step toward disentangling the graphical interface from the mathematics. Will need to revisit the cost of spawning threads in the single-point lift (I suspect it will cause undesireable latency in the UI).